Rename Tile to GridTile to reduce confusion with Tile Component
