+++ ShipDesigner/Assets/Workshop/Grid/GridController.cs
@@ -14,7 +14,7 @@ namespace Workshop.Grid
 		private int m_tileCountY; //number of tiles y
 		private Quad m_quad; // quad to render the grid
 		private Vector2 m_tileSize; // size of each time
-		private List<Tile> m_tiles; // collection of tiles in the grid
+		private List<GridTile> m_tiles; // collection of tiles in the grid
 
 		#endregion
 		#region Public Variables
@@ -48,7 +48,7 @@ namespace Workshop.Grid
 		/// </summary>
 		/// <param name="inputVector">Vector3 to compare to the tile locations</param>
 		/// <returns></returns>
-		public Tile GetTileByVector3(Vector3 input)
+		public GridTile GetTileByVector3(Vector3 input)
 		{
 			for (int i = 0; i <= m_tiles.Count; i++)
 			{
@@ -65,7 +65,7 @@ namespace Workshop.Grid
 
 		private void GenerateTileList()
 		{
-			m_tiles = new List<Tile>();
+			m_tiles = new List<GridTile>();
 
 			float width = m_quad.Vertices[2].x / m_quad.UVs[2].x;
 			float height = m_quad.Vertices[2].y / m_quad.UVs[2].y;
@@ -76,7 +76,7 @@ namespace Workshop.Grid
 				{
 					float pos_x = m_startLocation.x + (x * width);
 					float pos_y = m_startLocation.y + (y * height);
-					Tile tile = new Tile(pos_x, pos_y, m_tileSize);
+					GridTile tile = new GridTile(pos_x, pos_y, m_tileSize);
 					m_tiles.Add(tile);
 				}
 			}
